apps/workers/requirements.txt
# Created automatically by Cursor AI (2025-08-25)
# Python Workers for GIS and Modeling

# Core dependencies
celery==5.3.4
redis==5.0.1
nats-py==2.6.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0

# GIS and spatial data
geopandas==0.14.1
shapely==2.0.2
pyproj==3.6.1
fiona==1.9.5
rasterio==1.3.8
geopy==2.4.0

# Data processing
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4
scikit-learn==1.3.2

# Modeling and optimization
ortools==9.8.3296
pulp==2.7.0
networkx==3.2.1

# File processing
openpyxl==3.1.2
xlsxwriter==3.1.9
reportlab==4.0.7
pypdf2==3.0.1

# AWS/S3
boto3==1.34.0
minio==7.2.0

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
tenacity==8.2.3
pydantic==2.5.0
click==8.1.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
